# services/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Headway API",
    "description": "Serves computed transit metrics with full lineage, the DQ resolution workflow, and the audited certification action. This API never computes a reported figure; it serves what the calculation library produced, joined to its provenance. Reported values are JSON strings (exact NUMERIC, never float). Auth: local accounts (ADR-0011); the native OIDC relying party is the next increment behind the same {sub, username, role} claim set.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/settings/{setting_key}": {
      "put": {
        "tags": [
          "settings"
        ],
        "summary": "Update Setting",
        "description": "Change one policy setting (certifying official only). The old and new\nvalues land together in the audit trail; the change and its audit event\ncommit in one transaction.",
        "operationId": "update_setting_settings__setting_key__put",
        "parameters": [
          {
            "name": "setting_key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Setting Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSettingRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateSettingResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "UpdateSettingRequest": {
        "properties": {
          "value": {
            "type": "string",
            "minLength": 1,
            "title": "Value"
          }
        },
        "type": "object",
        "required": [
          "value"
        ],
        "title": "UpdateSettingRequest"
      },
      "UpdateSettingResponse": {
        "properties": {
          "setting_key": {
            "type": "string",
            "title": "Setting Key"
          },
          "setting_value": {
            "type": "string",
            "title": "Setting Value"
          },
          "value_type": {
            "type": "string",
            "title": "Value Type"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "updated_by": {
            "type": "string",
            "title": "Updated By"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "audit_event_id": {
            "type": "integer",
            "title": "Audit Event Id"
          }
        },
        "type": "object",
        "required": [
          "setting_key",
          "setting_value",
          "value_type",
          "description",
          "updated_by",
          "updated_at",
          "audit_event_id"
        ],
        "title": "UpdateSettingResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}